requirements.txt
flask==3.0.3
requests==2.32.3
gunicorn==23.0.0
mysql-connector-python==9.0.0
